LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/return-value/traverseTo-detach-cross-document-expected.txt



FAIL traverseTo() promise rejections when detaching an iframe inside onnavigate (cross-document) assert_equals: expected 1 but got 0
